% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_testing.R
\name{differential_test}
\alias{differential_test}
\title{Differential testing of abundant/expression of a given feature for a variable with two levels}
\usage{
differential_test(
  MAE,
  tax_level,
  variable = c(),
  min_num_filter = 5,
  padj_cutoff = 0.05,
  sig_only = c("Yes", "No")
)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{tax_level}{The classification level used for feature}

\item{variable}{Compare groups by binary variable e.g. 'Disease State'}

\item{min_num_filter}{Filtering of minimal number of counts. Defaults to 5}

\item{padj_cutoff}{Cutoff for p value adjustment, using BH. Defaults to 0.05}

\item{sig_only}{Show only significant values in plot}
}
\value{
A plotly object, a ggplot object for pdf saving, and a table with statistics
}
\description{
Differential testing of abundant/expression of a given feature for a variable with two levels
}
\examples{
data_dir = system.file('inst/extdata/Disease_Challenge.rds', package = 'QuickFixR')
df <- readRDS(data_dir)
p <- differential_test(df,
                    tax_level='Genus',
                    variable='Disease_State',
                    sig_only="No")
p

}
