% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Test Function For Package}
\usage{
hello(x)
}
\arguments{
\item{x}{The Name Of The Person To Say Hello.}
}
\value{
The output from \code{\link{print}}
}
\description{
Test Function For Package
}
\examples{
hello("Mahmood Jan")
\dontrun{
hello("Ziba Jan")
}
}
