% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_cor.R
\name{lme_cor}
\alias{lme_cor}
\title{Linear Mixed Effect Models with random variable}
\usage{
lme_cor(
  MAE,
  tax_level,
  feature,
  exp_var,
  random_var,
  datatype = c("logcpm", "relabu", "counts"),
  tolerance
)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{tax_level}{The classification level used for feature}

\item{feature}{A given feature from count data, which should be correlated with variable}

\item{random_var}{Random variable, which should be taken into account for correlation, eg. 'Location' or 'batch effect'.}

\item{datatype}{Select datatype, like relative abundance (relabu), counts, or logcpm}

\item{tolerance}{Tolerance of dispersion for correlation adjustment}

\item{variable}{Numeric variable for correlation, eg. 'weight'}
}
\value{
A plotly object, a ggplot object for pdf saving, and a table with statistics
}
\description{
Linear Mixed Effect Models with random variable
}
\examples{
data_dir = system.file('inst/extdata/NWS_MAG_Profiling.rds', package = 'QuickFixR')
df <- readRDS(data_dir)
p <- lme_cor(df,
                    tax_level='Genus',
                    feature = 'Mycoplasma'
                    variable='Weight_g',
                    random_var="Location",
                    datatype = "relabu")
p

}
