% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_test.R
\name{prevalence_test}
\alias{prevalence_test}
\title{Prevalence test for a variable of a given feature}
\usage{
prevalence_test(
  MAE,
  variable,
  tax_level,
  prev_threshold = 0.05,
  filtering_cutof = 0.001,
  datatype = c("logcpm", "relabu", "counts")
)
}
\arguments{
\item{MAE}{A multi-assay experiment object}

\item{variable}{Compare groups by binary variable e.g. 'Disease State'}

\item{tax_level}{The classification level used for feature}

\item{prev_threshold}{Threshold for prevalence testing. Defaults to 0.05}

\item{filtering_cutof}{Cutoff for filtering prior the test. Defaults to 0.001}

\item{datatype}{counts, relative abundance,logcpm}
}
\value{
A plotly object and a table of statistics
}
\description{
Prevalence test for a variable of a given feature
}
\examples{
data_dir = system.file('inst/extdata/Disease_Challenge.rds', package = 'QuickFixR')
df <- readRDS(data_dir)
p <- prevalence_test(df,
                    tax_level='Genus',
                    variable='Disease_State',
                    datatype='relabu')
p

}
