% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickFix.R
\name{QuickFix}
\alias{QuickFix}
\title{Run QucikFixR shiny app}
\usage{
QuickFix(dev = FALSE)
}
\arguments{
\item{dev}{Run the applicaiton in developer mode}
}
\value{
The shiny app will open
}
\description{
Run QucikFixR shiny app
}
\examples{
\dontrun{
QuickFix()
}
}
